🏗️ Architect: Modularized equipment.py into a package
Refactored the 755-line `apts/equipment.py` 'God Object' into a modular package `apts/equipment/`.
The `Equipment` class was decomposed using a MixIn pattern to separate core graph logic from plotting and data aggregation:
- Core `Equipment` logic and graph management moved to `base.py`.
- Visualization and plotting methods migrated to `EquipmentPlottingMixIn` in `plotting.py`.
- Helper class `MatplotlibSVGWrapper` isolated in `models.py`.
- Public API maintained via `__init__.py` to ensure backward compatibility.

This improvement follows the Single Responsibility Principle and improves maintainability of the equipment graph and reporting logic.
